+++ src/test/java/kitchenpos/order/OrderServiceTest.java
@@ -15,11 +15,11 @@ import kitchenpos.menu.dao.MenuRepository;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.dao.OrderDao;
 import kitchenpos.order.dao.OrderLineItemDao;
-import kitchenpos.order.dao.OrderTableRepository;
+import kitchenpos.table.dao.OrderTableRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
