+++ src/test/java/kitchenpos/table/TableGroupServiceTest.java
@@ -9,8 +9,8 @@ import static org.mockito.BDDMockito.given;
 import java.util.Arrays;
 import java.util.Collections;
 import kitchenpos.order.dao.OrderDao;
-import kitchenpos.order.dao.OrderTableRepository;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.dao.OrderTableRepository;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.application.TableGroupService;
 import kitchenpos.table.dao.TableGroupRepository;
 import kitchenpos.table.domain.TableGroup;
